requirements.txt
django==2.1.1
djangorestframework==3.8.2
psycopg2-binary==2.7.5
lxml==4.2.4
factory-boy==2.11.1
tqdm==4.25.0